Move IntervalMixin to separate file
