+++ posthog/models/filters/mixins/stickiness.py
@@ -4,7 +4,8 @@ from typing import Callable, Optional, Union
 from rest_framework.exceptions import ValidationError
 
 from posthog.constants import DATE_FROM, DATE_TO, STICKINESS_DAYS
-from posthog.models.filters.mixins.common import BaseParamMixin, DateMixin, IntervalMixin
+from posthog.models.filters.mixins.common import BaseParamMixin, DateMixin
+from posthog.models.filters.mixins.interval import IntervalMixin
 from posthog.models.filters.mixins.utils import cached_property, include_dict
 from posthog.models.team import Team
 from posthog.utils import relative_date_parse
