+++ posthog/models/filters/test/test_filter.py
@@ -1,7 +1,6 @@
 import json
 from typing import Callable, Optional
 
-import pytest
 from dateutil.relativedelta import relativedelta
 from django.db.models import Q
 from django.utils import timezone
@@ -535,35 +534,3 @@ class TestDateFilterQ(BaseTest):
             one_week_ago = timezone.now().replace(hour=0, minute=0, second=0, microsecond=0) - relativedelta(days=7)
             date_filter_query = filter.date_filter_Q
             self.assertEqual(date_filter_query, Q(timestamp__gte=one_week_ago, timestamp__lte=timezone.now()))
-
-
-@pytest.mark.parametrize(
-    "filter,expected_interval",
-    [
-        (Filter(data={"interval": "hour"}), "hour"),
-        (Filter(data={"interval": "day"}), "day"),
-        (Filter(data={"interval": "week"}), "week"),
-        (Filter(data={"interval": "month"}), "month"),
-        # Downcasing
-        (Filter(data={"interval": "HoUR"}), "hour"),
-        # Blank filter
-        (Filter(data={"events": []}), "day"),
-        # Legacy support - translate minutes to hours!
-        (Filter(data={"interval": "minute"}), "hour"),
-    ],
-)
-def test_filter_interval_success(filter, expected_interval):
-    assert filter.interval == expected_interval
-    assert filter.interval_to_dict() == {"interval": expected_interval}
-
-
-@pytest.mark.parametrize(
-    "filter,expected_error_message",
-    [
-        (Filter(data={"interval": "foo"}), "Interval foo does not belong to SUPPORTED_INTERVAL_TYPES!"),
-        (Filter(data={"interval": 123}), "Interval must be a string!"),
-    ],
-)
-def test_filter_interval_errors(filter, expected_error_message):
-    with pytest.raises(ValueError, match=expected_error_message) as error:
-        filter.interval
